Separate FakeRemoveRemote into a separate file
